zfs_arc plugin: Use the kstat utility functions.
This is nearly a rewrite, since the plugin was in such a poor state. It
removed the ugly hack of typesetting stuff to kstat_t in order to
preserve the signature of get_zfs_value() and replaces it with
zfs_get_{derive,gauge}() which are OS specific. The za_read() function
is now OS agnostic.